test bench for data memory
I created a test bench for the data memory module. I pre-loaded the memory with three different values for testing the loading instructions and I believe it's working fine. I also tested the store instructions but there still seems to be some issues with SB. There are detailed test cases for writing and reading each byte in the TB if you want to check the waveform.